// frontend/src/api/openapi/models/SingleDocSingleUserAnnotationsExportJobParams.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type SingleDocSingleUserAnnotationsExportJobParams = {
  /**
   * The ID of the Project to export from
   */
  project_id: number;
  export_job_type: any;
  /**
   * The ID of the SDocument to get the data from.
   */
  sdoc_id: number;
  /**
   * The ID of the User to get the data from.
   */
  user_id: number;
};